+++ alembic/env.py
@@ -16,7 +16,7 @@ if config.config_file_name is not None:
 
 # add your model's MetaData object here
 # for 'autogenerate' support
-from backend.app.models import Base
+from backend.app.schemas import Base
 target_metadata = Base.metadata
 # target_metadata = None
